hr_rag/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
langchain>=1.0.0
langchain-google-genai>=2.0.0
langchain-postgres>=0.0.12
langchain-community>=0.3.0
psycopg[binary]>=3.2.0
pypdf>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
